Created sharpness graphs
- Added color vars that match ingame hexcodes
- Dark mode compatible
