+++ src/routes/damage/weapon/weapon-info/index.ts
@@ -1,0 +1,1 @@
+export { WeaponInfo } from './weapon-info';
