+++ src/routes/damage/weapon/weapon-info/sharpness-graph.tsx
@@ -1,0 +1,21 @@
+interface SharpnessGraphProps {
+  sharpness: number[];
+}
+
+/**
+ * Represents a weapon's sharpness as a colored bar
+ */
+export function SharpnessGraph({ sharpness }: SharpnessGraphProps) {
+  return (
+    <div className="weapon-info__sharpness">
+      {sharpness.map((sharpnessLength, index) => (
+        <div
+          className={`sharpness-cell sharp${index}`}
+          style={{
+            width: `${sharpnessLength * 4}px`
+          }}
+        />
+      ))}
+    </div>
+  );
+}
