+++ src/routes/damage/weapon/weapon.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
-import { WeaponInfo } from './weapon-info';
 import { WeaponArgReducerActions, WeaponArgsState } from './weapon-reducer';
 import { WeaponSelectors } from './weapon-selectors';
-import { Weapons } from 'mh3-data';
 
 export interface WeaponProps {
   weaponArgs: WeaponArgsState;
@@ -13,11 +11,6 @@ export interface WeaponProps {
  * Top-level container for all info pertaining to the weapon
  */
 export function Weapon({ weaponArgs, dispatchWeaponArgs }: WeaponProps) {
-  const selectedWeapon = React.useMemo(
-    () => Weapons.getWeapon(weaponArgs.weaponClass, weaponArgs.weaponId),
-    [weaponArgs.weaponClass, weaponArgs.weaponId]
-  );
-
   return (
     <div className="weapon">
       <h3>Weapon</h3>
@@ -25,7 +18,6 @@ export function Weapon({ weaponArgs, dispatchWeaponArgs }: WeaponProps) {
         weaponArgs={weaponArgs}
         dispatchWeaponArgs={dispatchWeaponArgs}
       />
-      <WeaponInfo weapon={selectedWeapon} />
     </div>
   );
 }
