+++ src/sass/abstracts/variables.scss
@@ -8,3 +8,11 @@ $media-tablet-width: 959px;
 
 // Colors
 $selection-color: $pt-intent-primary;
+
+$sharpness-red: #c50f3a;
+$sharpness-orange: #e85218;
+$sharpness-yellow: #e9be32;
+$sharpness-green: #64df00;
+$sharpness-blue: #3068ee;
+$sharpness-white: #f0f0f0;
+$sharpness-purple: #f088f0;
