+++ src/sass/pages/damage/weapon.scss
@@ -1,3 +1,5 @@
+@import '../../abstracts/';
+
 .weapon {
   &--selectors {
     display: flex;
@@ -5,6 +7,55 @@
     gap: 1em;
   }
 }
+
+.bp5-dark .weapon-info__sharpness {
+  border: 1px solid white;
+}
+
+.weapon-info {
+  @include flex-container();
+
+  &__sharpness {
+    border: 1px solid black;
+    background-color: #333333;
+    height: 1.2em;
+    width: 210px;
+    padding: 1px;
+    margin-bottom: 0.5em;
+    .sharpness-cell {
+      height: 100%;
+      float: left;
+    }
+    .sharp0 {
+      background-color: $sharpness-red;
+    }
+    .sharp1 {
+      background-color: $sharpness-orange;
+    }
+    .sharp2 {
+      background-color: $sharpness-yellow;
+    }
+    .sharp3 {
+      background-color: $sharpness-green;
+    }
+    .sharp4 {
+      background-color: $sharpness-blue;
+    }
+    .sharp5 {
+      background-color: $sharpness-white;
+    }
+    .sharp6 {
+      background-color: $sharpness-purple;
+    }
+  }
+
+  &__stats {
+    label {
+      font-weight: bold;
+    }
+  }
+}
+
 .unique-weapon-selectors {
   display: flex;
   flex-direction: row;
